+++ Fun.Blazor.Tests/AdaptiveTests.fs
@@ -80,20 +80,3 @@ let ``html adaptive complex condition tests`` () =
 
     store2.Publish 1
     result.MarkupMatches("<div>s1=2</div>")
-
-
-type DemoForm = { Name: string; Age: int }    
-
-[<Fact>]
-let ``adaptive form state tests`` () =
-    let data = { Name = "n1"; Age = 20 }
-    use form = new AdaptiveForm<DemoForm, string>(data)
-    
-    form.UseFieldSetter(fun x -> x.Age)(21)
-    Assert.Equal(21, form.GetFieldValue(fun x -> x.Age))
-    Assert.Equal(true, form.UseHasChanges().Value)
-
-    form.SetValue(data)
-    Assert.Equal(data, form.GetValue())
-    Assert.Equal(false, form.UseHasChanges().Value)
-